Run Telemetry minidump tests on Android
We've had a number of issues with Telemetry minidump symbolization on
Android that hasn't been caught in part because the suite they belong
to, telemetry_perf_unittests, is not currently run on Android due to its
long runtime. This adds two invocations of the suite on the main Android
builders (one for ChromePublic.apk, one for MonochromePublic.apk), but
restricts the run to only the minidump tests, which only take several
minutes to run.

Bug: 1073267
Change-Id: I279820d67925bdbb12b3fe0ba666544df7b895e9
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2163398
Commit-Queue: Brian Sheedy <bsheedy@chromium.org>
Auto-Submit: Brian Sheedy <bsheedy@chromium.org>
Reviewed-by: Haiyang Pan <hypan@google.com>
Cr-Commit-Position: refs/heads/master@{#765848}