No GPU kill if the watchdog thread itself is slow
If it takes 15 more seconds than the expected time between two
OnWatchdogTimeout() calls, the system is considered slow and it's not a
GPU hang. The GPU process will not be killed this time.

This is ported from the GPU watchdog V1 which uses 2* timeout to
measure if it's too long for the watchdog thread. V2 uses +15 seconds
instead.

Bug: 949839
Change-Id: I191b3fce850b3a9f018334770b7aaf2c3f845bed
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2182401
Reviewed-by: Zhenyao Mo <zmo@chromium.org>
Commit-Queue: Maggie Chen <magchen@chromium.org>
Cr-Commit-Position: refs/heads/master@{#765862}