[fuchsia] Fix initialization with no message loop
Some test suites initialize us without a message loop. To bring up those
tests, we either have to allow that or make structural changes to the
tests. This does the former using the same approach as other ports of
delaying initialization that requires a message loop to creation of
a window.

Bug: 1075803

Change-Id: Iea57556df85db887a70a72a093baa9cb27d63350
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2173464
Commit-Queue: Michael Spang <spang@chromium.org>
Reviewed-by: Sergey Ulanov <sergeyu@chromium.org>
Cr-Commit-Position: refs/heads/master@{#765863}