Improve RAW file special cases when flipping chrome://flags#media-app
IsRawFile() wasn't doing a case-insensitive extension comparison, so
fix that.

But attempting to get the test to _fail_ without the fix exposed a gap
in test coverage. The "change detector" test doesn't use real files, so
it can't take advantage of "mime sniffing" used in the mime-type
determination within the extension file handling machinery. This meant
that RAW files (which get a "sniffed" MIME type of image/tiff) could be
directed to the Media App as a non-default user choice, which we don't
want.

Fixing that was necessary to get proper coverage of the `IsRawFile()`
check in tests to ensure the case-insensitivity was working correctly.
(Without that fix IsRawFile was skipped because the Media App was not
an eligible task at all, since the file had no mime type).

Test cases that would need mime-sniffing are given an explicit mime type
determined experimentally with the sniffer.

Bug: 1078231
Test: FileTasksBrowserTestWithMediaApp.DefaultHandlerChangeDetector
Change-Id: I3193d87899e1b798ef46fe2b0f9460162e9fdeaa
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2181933
Reviewed-by: Luciano Pacheco <lucmult@chromium.org>
Commit-Queue: Trent Apted <tapted@chromium.org>
Cr-Commit-Position: refs/heads/master@{#765867}