ime: Add fake shared library to test native typing using DecoderEngine.
Right now our native typing browser tests are using InputEngine,
which only works for rule-based typing. We want to have tests for
other input methods too, which requires DecoderEngine.

Unfortunately, DecoderEngine needs to load a shared library from the
filesystem. So when we run our tests with DecoderEngine, we get an error
that it can't find the shared library.

We create a simple fake shared library so that DecoderEngine can be
tested. We follow a similar pattern as [1].

Since DecoderEngine runs inside the IME sandbox, we also need to
disable the sandbox during our tests.

[1] https://cs.chromium.org/chromium/src/chrome/test/BUILD.gn?q=verifier_test_dll_1&l=5228&dr=C

Bug: 1019541
Change-Id: Iefdab7d06c375dc2a8adf08dfd799da2d95f96bb
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2125915
Commit-Queue: Darren Shen <shend@chromium.org>
Reviewed-by: Leo Zhang <googleo@chromium.org>
Reviewed-by: Keith Lee <keithlee@chromium.org>
Cr-Commit-Position: refs/heads/master@{#765879}