[iOS] Handle sub frame unsafe resources
Updates SafeBrowsingTabHelper::PolicyDecider::OnUrlQueryDecided() to
reload the last-committed NavigationItem if an unsafe resource is
detected in a subframe of that navigation.  Upon reloading, the main
frame request is immediately cancelled with a safe browsing error to
show the error page for the unsafe subframe resource.

Bug: 1064803
Change-Id: Ib4ba4d5d565276a222e23f2e388b5a6e2f57e896
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2176674
Commit-Queue: Kurt Horimoto <kkhorimoto@chromium.org>
Reviewed-by: Ali Juma <ajuma@chromium.org>
Cr-Commit-Position: refs/heads/master@{#765880}