[Launcher impressions] Call notifier from AppListView::SetState.
Currently we are updating the AppListNotifier of a view state change
when the transition animation is completed. This adds some latency
that's tricky to account for, especially on old devices. Instead, let's
move the update to AppListView::SetState, so it gets called before the
animation begins.

This information needs to travel from the AppListView to the
AppListViewDelegate. I couldn't find an existing method that could serve
for this kind of update, so I've added
AppListViewDelegate::OnViewStateChanged.

Bug: 1076270
Change-Id: I387b96934a11f97a57ac56473caae7da9c145b93
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2182093
Reviewed-by: Jenny Zhang <jennyz@chromium.org>
Commit-Queue: Tony Yeoman <tby@chromium.org>
Cr-Commit-Position: refs/heads/master@{#765901}