Redesign sync proxy datatype configuration
ProxyTabsDataTypeController was the last remnant of a controller that
behaved like the old Directory-based controllers, using
Directory-specific APIs for activation, and adhering to a slightly
different state machine due to ShouldLoadModelBeforeConfigure()
returning false.

In this patch, this controller aligns in behavior with
ModelTypeController, which allows simplifying the base
DataTypeController API and the corresponding calling sites.

To avoid directory-dependent primitives without a drastic redesign of
how proxy types are configured, dedicated APIs are introduced for proxy
types in ModelTypeConfigurer (replacing the Directory-centric ones) and
analogous layers within the sync engine (ModelTypeConnector; but here
the cleanup can only take place after Nigori USS is cleaned up).

Change-Id: I12058d7dc5b631b0963f4ffb9344a7c1f6f5df57
Bug: 1068097
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2179083
Commit-Queue: Mikel Astiz <mastiz@chromium.org>
Reviewed-by: Marc Treib <treib@chromium.org>
Cr-Commit-Position: refs/heads/master@{#765909}