CSP: parse 'upgrade-insecure-request'
We can parse it and it is easy to do. So let's do this.
The value parsed from the network process won't be used for now. Only
the one filled from the renderer process will. We might want to reverse
the network/renderer responsibility at some point in the future.

Bug: 1076894
Change-Id: I35e72fc699671958f90ce42ef6ed405a5a24943c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2164815
Commit-Queue: Arthur Sonzogni <arthursonzogni@chromium.org>
Auto-Submit: Arthur Sonzogni <arthursonzogni@chromium.org>
Reviewed-by: Mike West <mkwst@chromium.org>
Cr-Commit-Position: refs/heads/master@{#765927}