Remove the IncognitoDataTestUtils#finishActivities
The IncognitoDataTestUtils#finishActivities method iterates over all
running activities and invoke the finish method on activity which is
of instance ChromeActivity.

The TestRules CustomTabActivityTestRule and ChromeActivityTestRule
already owns the lifecycle of the activities they created.

This CL removes the explicit call to the finish method for these
activities.

Bug: 1035770
Change-Id: Ie7b7a14b518cab48b32bb75eac4327302226622d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2184249
Reviewed-by: Ramin Halavati <rhalavati@chromium.org>
Reviewed-by: Peter Conn <peconn@chromium.org>
Reviewed-by: Rohit Agarwal <roagarwal@chromium.org>
Commit-Queue: Rohit Agarwal <roagarwal@chromium.org>
Cr-Commit-Position: refs/heads/master@{#765948}