bindings: Fix teardown of V8 GC callbacks
Prologue and epilogue callbacks were never removed. The callbacks
assume that ActiveScriptWrappableManager object is available. Since
this object gets cleared in WillBeDestroyed it is gone though during
teardown.

To maintain the invariant that the object is always available when V8
is connected, we properly remove the callbacks which should happen
anyways.

Bug: chromium:1078751, chromium:1056170
Change-Id: I4972d8072c9497bd5aabddc66e6a7f9ab5a476b4
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2184173
Reviewed-by: Kentaro Hara <haraken@chromium.org>
Commit-Queue: Michael Lippautz <mlippautz@chromium.org>
Cr-Commit-Position: refs/heads/master@{#765966}