[PM] Create GraphRegistered class.
This is a helper class that allows objects to be registered
with the Graph by TypeId, allowing the Graph to be used as a
rendezvous point for graph-associated singletons.

For example, it is common for UI thread instrumentation to want
to redirect to a GraphOwned object on the PM sequence. Currently
each such instance maintains its own logic to do this. This helper
class will make this much easier.

Change-Id: I18ab5de898c2cdbe62d73893415a6d1cfb57ee31
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2180762
Commit-Queue: Chris Hamilton <chrisha@chromium.org>
Reviewed-by: Sigurður Ásgeirsson <siggi@chromium.org>
Auto-Submit: Chris Hamilton <chrisha@chromium.org>
Cr-Commit-Position: refs/heads/master@{#766017}