Allow WebUIs to iframe web content
WebUIs were not allowed to iframe web content in the past, since we
could not properly isolate the web iframes from the WebUI. However, we
have now shipped some form of site isolation on all platforms and we
can isolate the web iframes into a separate process from the WebUI,
separately from whether full site isolation is enabled or not.
The exception to this rule for now is Android WebView, which does not
yet lock WebUIs processes to origin.

With WebUI embedding web content in iframe, the two will have to run in
the same BrowsingInstance, which is different from the current behavior
where we don't allow WebUI and web content to share a BrowsingInstance.

This CL removes the WebUINavigationThrottle, which was enforcing
that web iframes cannot be added to WebUIs or processes with WebUI
bindings. The regular process model is now responsible for ensuring
that navigations to WebUIs are properly isolated and we have added
explicit security checks to ensure that non-WebUI documents are not
committed in processes with WebUI bindings, so it is now safe to remove
the throttle, given the goal of allowing WebUI to iframe web content.

We still do not want to allow web content to iframe WebUIs, which
remains unchanged with this CL and is enfroced both by the renderer
process and browser process separately.

Bug: 1002276, 683418, 1035292
Change-Id: I54bee7f3a26043a2986259123930619107f6e268
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2007813
Commit-Queue: Nasko Oskov <nasko@chromium.org>
Reviewed-by: Bo <boliu@chromium.org>
Reviewed-by: Charlie Reis <creis@chromium.org>
Reviewed-by: Alex Moshchuk <alexmos@chromium.org>
Reviewed-by: Julian Pastarmov <pastarmovj@chromium.org>
Reviewed-by: Charlie Harrison <csharrison@chromium.org>
Cr-Commit-Position: refs/heads/master@{#766052}