Fix the crash issue when launching an uninstalled Crostini app.
When running a Crostini app binary which is not installed, and the app
info is not saved in AppService,
AppServiceAppWindowCrostiniTracker::GetShelfAppId can return the
correct app_id when the window is activated:
https://source.chromium.org/chromium/chromium/src/+/master:chrome/browser/ui/ash/launcher/app_service/app_service_app_window_crostini_tracker.cc;l=192
So the window is saved to InstanceRegistry, and saved to AppWindowBase.

However, when the window is closed,
AppServiceAppWindowCrostiniTracker::GetShelfAppId can't return the
app_id, and because the app is not installed, not saved in AppService,
we can't get the app_id from
AppServiceAppWindowLauncherController::GetShelfId:
https://source.chromium.org/chromium/chromium/src/+/master:chrome/browser/ui/ash/launcher/app_service/app_service_app_window_launcher_controller.cc;l=535
So the window is not destroyed from AppService, and AppWindowBase is
not destroyed as well, which causes the crash:
https://source.chromium.org/chromium/chromium/src/+/master:chrome/browser/ui/ash/launcher/app_service/app_service_app_window_launcher_controller.cc;l=228

In the previous Crostini app window code, when the window is destroyed,
it doesn't check the app id:
https://source.chromium.org/chromium/chromium/src/+/master:chrome/browser/ui/ash/launcher/crostini_app_window_shelf_controller.cc;l=297
So AppServiceAppWindowLauncherController::OnWindowDestroying should do
the similar thing as well. If we can't get the app_id from
AppServiceAppWindowCrostiniTracker::GetShelfAppId, we can check whether
the window is saved in InstanceRegistry. If yes, we can get the
shelf_id/app_id from InstanceRegistry, and delete the record from
InstanceRegistry, and remove the AppWindowBase.

BUG=1073937

Change-Id: I7ca84b02c9a17a96394459284f80d40b4bf3e913
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2185349
Reviewed-by: Xiyuan Xia <xiyuan@chromium.org>
Commit-Queue: Nancy Wang <nancylingwang@chromium.org>
Cr-Commit-Position: refs/heads/master@{#766076}