Move blink/public/mojom/frame/ .mojom files to the "mojom_core" target
As discussed on the mailing list [1], we're moving the frame.mojom from
the "mojom_platform" target into the "mojom_core" one, in preparation
for the migration of FrameHostMsg_RouteMessageEvent to mojo, which will
need having frame.mojom depend on messaging/transferable_message.mojom.

For consistency, though, this CL does not only move the frame.mojom
file but also all the .mojom files from frame/, that were previously
in "mojom_platform", into "mojo_core" and "frame" depending on which
of them depend or not on "mojom_platform", to prevent cycles. This
addresses a situation that was already a bit strange, in which the
"frame" target only had "find_in_page.mojom" because most of the other
.mojom files in the frame/ directory depended on "mojo_platform",
which is no longer the case with this CL.

[1] https://groups.google.com/a/chromium.org/d/msg/chromium-mojo/Sk30ujUCNPI/9zPqpqqqBAAJ

Bug: 1047383
Change-Id: I651094f6b8c3aaf0eb139406b383fe3183cab3cd
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2182378
Commit-Queue: Mario Sanchez Prada <mario@igalia.com>
Reviewed-by: Kentaro Hara <haraken@chromium.org>
Reviewed-by: Dave Tapuska <dtapuska@chromium.org>
Cr-Commit-Position: refs/heads/master@{#766099}