Extend expiration for password site isolation about:flags entry.
While this feature shipped to all users in M77, the launch only
applied to devices having at least 1900MB of RAM.  This flag allows
users of lower-end devices to still opt in to the Android mode of site
isolation, so it makes sense to keep it around for longer.

Note that there are various experiments in progress about
site-isolating more devices, including some of the lower-end devices.
Depending on the outcome of those, let's revisit whether this flag
still makes sense in a year or so.

Change-Id: Ibd02d7670032f376a06d04eaf2de4ffb8a35aa9c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2181471
Auto-Submit: Alex Moshchuk <alexmos@chromium.org>
Commit-Queue: Łukasz Anforowicz <lukasza@chromium.org>
Reviewed-by: Łukasz Anforowicz <lukasza@chromium.org>
Cr-Commit-Position: refs/heads/master@{#766109}