[fuchsia] Add GLSurfaceFuchsiaImagePipe
This implements CreateViewGLSurface on Fuchsia, so that we can run tests
that use that (e.g. anything that uses InProcessCommandBuffer).

It also adds an owned PlatformWindowSurface to InProcessCommandBuffer,
which is necessary because we split onscreen surfaces into two parts on
Fuchsia, a renderer agnostic part (PlatformWindowSurface) and a renderer
specific one (in this case GLSurface).

Bug: 1075803

Change-Id: I860195024216b0c1c82cc7a53607f21a8a25ea68
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2173708
Reviewed-by: Geoff Lang <geofflang@chromium.org>
Reviewed-by: Sergey Ulanov <sergeyu@chromium.org>
Reviewed-by: Jonathan Backer <backer@chromium.org>
Commit-Queue: Michael Spang <spang@chromium.org>
Cr-Commit-Position: refs/heads/master@{#766110}