Reland: "[PM] Create GraphRegistered class."
This is a helper class that allows objects to be registered
with the Graph by TypeId, allowing the Graph to be used as a
rendezvous point for graph-associated singletons.

For example, it is common for UI thread instrumentation to want
to redirect to a GraphOwned object on the PM sequence. Currently
each such instance maintains its own logic to do this. This helper
class will make this much easier.

This is a reland of

https://chromium.googlesource.com/chromium/src/+/95fba6ec8b32eb6f804e43e007ebee19314e0832

which was reverted here:

https://chromium.googlesource.com/chromium/src/+/bc0e14ed148365319343f24e0115a61c32f1298f

Failures were due to erroneous EXPECT_CHECK_DEATH instead of
EXPECT_DCHECK_DEATH statements, and the failing bots were release
builds without DCHECKs enabled.

Change-Id: I6477ac76f04e75378847d3b2428623905a5d690c

TBR=siggi@chromium.org

Change-Id: I6477ac76f04e75378847d3b2428623905a5d690c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2185389
Commit-Queue: Chris Hamilton <chrisha@chromium.org>
Reviewed-by: Chris Hamilton <chrisha@chromium.org>
Cr-Commit-Position: refs/heads/master@{#766116}