Add robolectric to android_deps.
Robolectric needs to be compiled, but shouldn't be labelled as
supports_android = true as that'll cause various deps to need to
be supports_android = true (such as guava) so made a new testCompile
and androidTestCompile gradle configs.

All licenses are MIT and apache 2.0, except for backport which is
CC O1.

Bug: 1060721
Change-Id: I541f643231bebb15b9be0ea9191912233a731774
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1972850
Commit-Queue: Peter Wen <wnwen@chromium.org>
Reviewed-by: Yaron Friedman <yfriedman@chromium.org>
Reviewed-by: benjamin joyce <bjoyce@chromium.org>
Reviewed-by: Ryan Sleevi <rsleevi@chromium.org>
Reviewed-by: Peter Wen <wnwen@chromium.org>
Reviewed-by: Andrew Grieve <agrieve@chromium.org>
Auto-Submit: benjamin joyce <bjoyce@chromium.org>
Cr-Commit-Position: refs/heads/master@{#766130}