[WebLayer] Move site_settings to components/browser_ui/site_settings
This CL moves the majority of //chrome/browser/site_settings to
//components/browser_ui/site_settings, leaving only Chrome-specific
classes behind. This is part of an effort to componentize site_settings
so it can be reused by WebLayer.

This CL leaves most of the tests in //chrome because they rely on the
SettingsActivity and specific chrome implementation of the site
settings clients. I felt that componentizing these tests would result
in removing enough of the chrome based scaffolding around the UI that
it would reduce the effectiveness of the tests.

Bug: 1058595
Change-Id: I5bee6e2a36e5e9efb5658b42ab11a532e2ee5853
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2169747
Commit-Queue: Robbie McElrath <rmcelrath@chromium.org>
Reviewed-by: Theresa  <twellington@chromium.org>
Reviewed-by: Finnur Thorarinsson <finnur@chromium.org>
Reviewed-by: Colin Blundell <blundell@chromium.org>
Reviewed-by: Bo <boliu@chromium.org>
Cr-Commit-Position: refs/heads/master@{#766134}