[iOS] Remove usages of MDC Snackbar in Activities
As part of the effort to support sharing an image in
the activity view, I found out that using the native
Copy activity will always raise a "Link copied"
snackbar message - even for images.

After chatting with UX, we figured that these snackbar
messages should simply be removed (as part of the grand
MDC removal principle), which is what this CL is
achieving.

I did not remove the usage in send_tab_to_self since
it was out of activity_services/ scope.

Bug: 1068606
Change-Id: Idd7a967edfcf9e4cd3b777b8625ec25ea24015ff
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2181313
Commit-Queue: Sebastien Lalancette <seblalancette@chromium.org>
Reviewed-by: Gauthier Ambard <gambard@chromium.org>
Reviewed-by: Jeffrey Cohen <jeffreycohen@chromium.org>
Cr-Commit-Position: refs/heads/master@{#766160}