Simplify ScreenMac
In ScreenMac there exists a cache of display::Displays
built from NSScreens. This cache started being used in
crrev.com/433463.

Some effort is made to minimize the number of times that
we update this cache (NSScreens change more than once in
rapid succession when the display configuration changes).
This gets complicated because we (1) want to delay
re-computing the cache, but (2) want to make sure it's
up-to-date when queried and (3) want to ensure that we
appropriately update DisplayObservers.

These goals are hard to satisfy. This patch gives up
on goal (1) and just re-computes the cache every time
the display configuration changes. This will do more
work than we did before, but only when plugging or
unplugging a monitor or changing display settings,
both of which are infrequent.

Bug: 1033866
Change-Id: I1d7d22c3a968d1563412d1d888c09f19d3105017
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2135125
Commit-Queue: ccameron <ccameron@chromium.org>
Reviewed-by: Elly Fong-Jones <ellyjones@chromium.org>
Cr-Commit-Position: refs/heads/master@{#766184}