WebXR - blink - stop using std::unique_ptr for optional return types
Currently, in blink code, we have been using std::unique_ptr in return
types from methods (and sometimes in arguments method accept) to express
that the result may be unavailable. This was due to the implementation
detail leak (blink::TransformationMatrix needs to be stored in
oilpan-heap-managed types via a std::unique_ptr to satisfy its alignment
requirements). This change replaces std::unique_ptr with base::Optional
for return types to communicate clearly that they are optional (as
opposed to sending an incorrect signal about memory ownership).

This is a mechanical change, there should be no functional changes.

Change-Id: I2c872b8c94a5f6a8f488bb9acf69762956ad0e64
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2185793
Reviewed-by: Brandon Jones <bajones@chromium.org>
Commit-Queue: Piotr Bialecki <bialpio@chromium.org>
Cr-Commit-Position: refs/heads/master@{#766193}