[IntentHandling] Add a flag for PWA persistence in intent picker.
Add a flag for showing the PWA persistence in the new intent picker. So
we can roll out the PWA persistence after the intent handling backend.
My plan is to release the exact same intent picker with the intent
handling backend first, then enable the PWA persistence, because
currently there is no easy way for user to manage the persistence.

BUG=853604

Change-Id: Ib6a2347ed1de37bf1729cfb281fb31fd49492537
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2181251
Reviewed-by: Dominick Ng <dominickn@chromium.org>
Reviewed-by: Scott Violet <sky@chromium.org>
Commit-Queue: Maggie Cai <mxcai@chromium.org>
Cr-Commit-Position: refs/heads/master@{#766232}