Avoid dumping output on success in generate_buildbot_json_unittest.
Some of the test cases intentionally set the JSON files to be different
than the generated output to ensure that they correctly fail consistency
checks which causes the output to be dumped even when the test passes.
The dump_on_failure function can be called separately so that if the
consistency check fails in an expected manner, the output won't be
dumped.

Change-Id: Ie06e8cbc72e3135e532d1ce2b043fb62a0c3336e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2185798
Reviewed-by: Stephen Martinis <martiniss@chromium.org>
Commit-Queue: Garrett Beaty <gbeaty@chromium.org>
Auto-Submit: Garrett Beaty <gbeaty@chromium.org>
Cr-Commit-Position: refs/heads/master@{#766237}