[ios] Configure ModalObservation only when banner is triggering modal
This change makes it so that InfobarBannerOverlayRequestCancelHandler
does not cancel itself when getting modal completion callbacks from
modals that it did not present. This adds an observer to
InfobarOverlayRequestInserter for banner cancel handlers to be notified
when requests are inserted.

InfobarOverlayRequestInserter's API is modified to accept InsertParams
to help specify the source the request insertion. AddOverlayRequest()
is removed is it is unnecessary at this point.

Bug: 1071914
Change-Id: I7806835683bbef08b833abad8dd676d57fb91e0f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2176419
Commit-Queue: Chris Lu <thegreenfrog@chromium.org>
Reviewed-by: Kurt Horimoto <kkhorimoto@chromium.org>
Cr-Commit-Position: refs/heads/master@{#766244}