Add mapping of payment method names to the id of the apps
EnableMethodManifestUrlForSupportedApps() in manifest_verifier.cc
currently marks ALL apps from an app_origin as enabled
for a payment method if "supported_origin_strings"
of the payment method manifest contains the app origin.
This is not always correct.

For example, given the following setup:
1) https://alicepay.com/ hosts two apps:
  - /app1 supports payment method "foo.com"
  - /app2 supports payment method "bar.com"
2) Both "foo.com" and "bar.com" are in specified in a given
   PaymentRequest
3) The payment method manifest at "foo.com" has "alicepay.com"
   in supported_origin_strings
4) The payment method manifest at "bar.com" does NOT have
   "alicepay.com" in supported_origin_strings without this change,
   both /app1 and /app2 will be marked as enabled with "foo.com"
   as the enabled method.

With this change, only /app1 will be enabled and /app2 will be
removed in ManifestVerifier::RemoveInvalidPaymentApps().

Bug: none

Change-Id: I5d5811f9a9a921e30c9bd1de611a7f3c0db07fdc
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2174249
Commit-Queue: DongJun Kim <djmix.kim@samsung.com>
Reviewed-by: Danyao Wang <danyao@chromium.org>
Cr-Commit-Position: refs/heads/master@{#766292}