device/fido: tighten up PublicKey.
PublicKey was an interface, implemented by OpaquePublicKey (used by
CTAP2) and P256PublicKey (used by U2F). But we only need to store the
CBOR bytes (and, in the future, the algorithm and DER bytes too). Thus
PublicKey can just be a container for a few members and the different
logic is only needed when parsing.

Thus this change makes PublicKey a concrete type that stores what we
need and P256PublicKey becomes a struct/namespace for functions for
creating PublicKey objects from P-256 public keys.

Additionally, apply sanity checks in U2F and CTAP2 paths that ensure,
for example, that the public key is on the curve.

Change-Id: I0e952200f10195e675bd2841100e51bd215b6d51
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2182931
Commit-Queue: Adam Langley <agl@chromium.org>
Auto-Submit: Adam Langley <agl@chromium.org>
Reviewed-by: Martin Kreichgauer <martinkr@google.com>
Cr-Commit-Position: refs/heads/master@{#766297}